+++ src/page/slim-page/components/header-actions.tsx
@@ -1,16 +1,16 @@
 import Button from "@material-ui/core/Button";
 import React, { forwardRef } from "react";
 import classNames from "classnames";
-import styles from "../common-page-v2.module.scss";
+import styles from "../slim-page.module.scss";
 import { ClickableItem } from "../../../types/dropdown";
-import { CommonPageV2Classes, HeaderAction } from "../../../types/card";
+import { SlimPageClasses, HeaderAction } from "../../../types/card";
 import { LinkButton } from "../../../link";
 import { SecondaryActionsMenuButton } from "../../../dev";
 
 type Props = {
   primaryActions?: HeaderAction[];
   secondaryActions?: ClickableItem[];
-  classes?: CommonPageV2Classes;
+  classes?: SlimPageClasses;
 };
 
 export const HeaderActions = forwardRef<any, Props>(