+++ src/page/slim-page/slim-tabbed-page.tsx
@@ -1,10 +1,10 @@
 import React, { ReactNode } from "react";
 import { ClickableItem } from "../../types/dropdown";
-import { CommonPageV2 } from "./common-page-v2";
+import { SlimPage } from "./slim-page";
 import {
-  CommonTabbedPageV2Classes,
+  SlimTabbedPageClasses,
   HeaderAction,
-  PageConfigV2,
+  SlimPageConfig,
 } from "../../types/card";
 import { TabProps } from "@material-ui/core/Tab";
 import { TabsComponent } from "./components/tabs-component";
@@ -14,7 +14,7 @@ type Props = {
    * The array of objects used to define the tabs displayed in the header, and their corresponding pages.
    * The `key` property is used to determine which tab is selected.
    */
-  pageConfigs: PageConfigV2[];
+  pageConfigs: SlimPageConfig[];
   /** Takes the handler used to switch between tabs, based on the tab's value */
   onChangeActiveTab: (x0: string) => any;
   /** The value of the tab header. Used to define which tab is open */
@@ -49,7 +49,7 @@ type Props = {
    *
    *  - `content` - (applied to the container around the content)
    */
-  classes?: CommonTabbedPageV2Classes;
+  classes?: SlimTabbedPageClasses;
   /** Takes a `node` to show on the page */
   children?: ReactNode;
   /** Primary actions to display on the header */
@@ -65,12 +65,12 @@ type Props = {
 };
 
 /**
- * `CommonTabbedPage` provides a component for a page with a flush card header,
+ * `SlimTabbedPage` provides a component for a page with a flush card header,
  * a tab bar, and an optional side menu. Note: if you are using this component
  * in a new environment, we recommend building a new child component for your
  * environment; take a look at the `TabbedLimsPage` component as an example.
  */
-export const CommonTabbedPageV2: React.FC<Props> = props => {
+export const SlimTabbedPage: React.FC<Props> = props => {
   const {
     classes = {},
     activeTab,
@@ -96,7 +96,7 @@ export const CommonTabbedPageV2: React.FC<Props> = props => {
   }
   const { tabs, tab, ...commonPageClasses } = classes;
   return (
-    <CommonPageV2
+    <SlimPage
       {...commonPageProps}
       classes={{ ...commonPageClasses }}
       isLoading={isLoading}
@@ -116,6 +116,6 @@ export const CommonTabbedPageV2: React.FC<Props> = props => {
     >
       {PageContent}
       {children}
-    </CommonPageV2>
+    </SlimPage>
   );
 };